refactor: moved out deploy results
